# specs/001-georectification/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ALPROJ GUI Backend API
  version: 1.0.0
  description: |
    景観写真オルソ化アプリのバックエンド API。
    フロントエンド（Svelte）から呼び出される。
servers:
  - url: http://localhost:{port}
    description: ローカルサイドカーサーバー
    variables:
      port:
        default: "8765"
...
paths:
...
  /api/files/raster/info:
    post:
      tags: [files]
      summary: ラスタファイル情報取得
      operationId: getRasterInfo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
      responses:
        "200":
          description: ファイル情報
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RasterFile"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/files/image/info:
    post:
      tags: [files]
      summary: 画像ファイル情報取得
      operationId: getImageInfo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
      responses:
        "200":
          description: ファイル情報
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageFile"
  /api/files/raster/thumbnail:
    post:
      tags: [files]
      summary: ラスタサムネイル取得
      operationId: getRasterThumbnail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                max_size:
                  type: integer
                  default: 512
      responses:
        "200":
          description: サムネイル画像
          content:
            image/png:
              schema:
                type: string
                format: binary

  # === Georectify ===
...
components:
  schemas:
...
    RasterFile:
      type: object
      required: [path, crs, bounds, resolution, size]
      properties:
        path:
          type: string
        crs:
          type: string
          example: "EPSG:6690"
        bounds:
          type: array
          items:
            type: number
          minItems: 4
          maxItems: 4
          description: "[xmin, ymin, xmax, ymax]"
        resolution:
          type: array
          items:
            type: number
          minItems: 2
          maxItems: 2
        size:
          type: array
          items:
            type: integer
          minItems: 2
          maxItems: 2
          description: "[width, height]"
    ImageFile:
      type: object
      required: [path, size]
      properties:
        path:
          type: string
        size:
          type: array
          items:
            type: integer
          minItems: 2
          maxItems: 2
        exif:
          $ref: "#/components/schemas/ExifData"
    ExifData:
      type: object
      properties:
        datetime:
          type: string
          format: date-time
        gps_lat:
          type: number
        gps_lon:
          type: number
        gps_alt:
          type: number
        focal_length:
          type: number
        camera_model:
          type: string

    # === Camera Params ===
...
  responses:
...
    BadRequest:
      description: リクエストが不正です
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              detail:
                type: string